+++ client/src/screens/hub/HubMainScreen/components/Panel/components/ParallelCoordinatesChart/ParallelCoordinatesChart.js
@@ -983,14 +983,14 @@ function ParallelCoordinatesChart(props) {
     return () => {
       window.cancelAnimationFrame(renderChart);
     };
-  }, [traceList, props.width, props.height]);
+  }, [traceList, chart, props.width, props.height]);
 
-  useEffect(() => {
-    window.requestAnimationFrame(renderData);
-    return () => {
-      window.cancelAnimationFrame(renderData);
-    };
-  }, [chart]);
+  // useEffect(() => {
+  //   window.requestAnimationFrame(renderData);
+  //   return () => {
+  //     window.cancelAnimationFrame(renderData);
+  //   };
+  // }, [chart]);
 
   return (
     <div className='ParallelCoordinatesChart' ref={parentRef}>